requirements.txt
flask>=3.0
requests>=2.31
pycryptodome>=3.20
jinja2>=3.1
supabase>=2.0
gunicorn>=22.0
